refactor(ui): extract shared zip helpers for db/logs export
Collapse the duplicated zip-filename and zip-writer logic shared by the
database and logs export features into src/main/ui/zip.ts:
buildTimestampedZipName, ensureZipExtension, createZipWithFiles. The
single-file DB path now uses the multi-file writer. Adds zip.test.ts
covering the name/extension helpers (previously untested).
