+++ src/main/ui/database-export.ts
@@ -3,8 +3,8 @@ import fs from 'node:fs'
 import * as fsPromises from 'node:fs/promises'
 import os from 'node:os'
 import path from 'node:path'
-import * as yazl from 'yazl'
 import log from '../logger'
+import { buildTimestampedZipName, createZipWithFiles, ensureZipExtension } from './zip'
 import type { DatabaseExportResult } from '../../shared/types'
 
 export interface DatabaseExportStorage {
@@ -17,46 +17,6 @@ interface ExportDatabaseZipOptions {
   parentWindow?: BrowserWindow | null
 }
 
-function pad2(v: number): string {
-  return String(v).padStart(2, '0')
-}
-
-function buildDefaultDatabaseExportFilename(now = new Date()): string {
-  const timestamp = `${now.getFullYear()}${pad2(now.getMonth() + 1)}${pad2(now.getDate())}-${pad2(
-    now.getHours(),
-  )}${pad2(now.getMinutes())}${pad2(now.getSeconds())}`
-  return `memorylane-db-export-${timestamp}.zip`
-}
-
-function ensureZipExtension(filePath: string): string {
-  return filePath.toLowerCase().endsWith('.zip') ? filePath : `${filePath}.zip`
-}
-
-async function createZipWithSingleFile(
-  inputPath: string,
-  entryName: string,
-  outputZipPath: string,
-): Promise<void> {
-  await fsPromises.mkdir(path.dirname(outputZipPath), { recursive: true })
-
-  await new Promise<void>((resolve, reject) => {
-    const zipFile = new yazl.ZipFile()
-    const output = fs.createWriteStream(outputZipPath)
-
-    const onError = (error: unknown): void => {
-      reject(error instanceof Error ? error : new Error(String(error)))
-    }
-
-    output.once('error', onError)
-    zipFile.outputStream.once('error', onError)
-    output.once('close', resolve)
-
-    zipFile.outputStream.pipe(output)
-    zipFile.addFile(inputPath, entryName)
-    zipFile.end()
-  })
-}
-
 export async function exportDatabaseZip({
   storage,
   parentWindow = null,
@@ -67,7 +27,7 @@ export async function exportDatabaseZip({
   try {
     const defaultPath = path.join(
       app.getPath('documents'),
-      buildDefaultDatabaseExportFilename(new Date()),
+      buildTimestampedZipName('memorylane-db-export'),
     )
     const saveResult = await dialog.showSaveDialog(parentWindow ?? undefined, {
       title: 'Export Database ZIP',
@@ -86,7 +46,7 @@ export async function exportDatabaseZip({
     const backupPath = path.join(tempDir, dbBasename)
 
     await storage.backupToFile(backupPath)
-    await createZipWithSingleFile(backupPath, dbBasename, outputPath)
+    await createZipWithFiles([backupPath], outputPath)
 
     return { success: true, outputPath }
   } catch (error) {
