+++ src/main/ui/logs-export.ts
@@ -2,27 +2,12 @@ import { app, dialog, type BrowserWindow } from 'electron'
 import fs from 'node:fs'
 import * as fsPromises from 'node:fs/promises'
 import path from 'node:path'
-import * as yazl from 'yazl'
 import log from '../logger'
+import { buildTimestampedZipName, createZipWithFiles, ensureZipExtension } from './zip'
 import type { DatabaseExportResult } from '../../shared/types'
 
 interface ExportLogsZipOptions {
   parentWindow?: BrowserWindow | null
-}
-
-function pad2(v: number): string {
-  return String(v).padStart(2, '0')
-}
-
-function buildDefaultLogsExportFilename(now = new Date()): string {
-  const timestamp = `${now.getFullYear()}${pad2(now.getMonth() + 1)}${pad2(now.getDate())}-${pad2(
-    now.getHours(),
-  )}${pad2(now.getMinutes())}${pad2(now.getSeconds())}`
-  return `memorylane-logs-export-${timestamp}.zip`
-}
-
-function ensureZipExtension(filePath: string): string {
-  return filePath.toLowerCase().endsWith('.zip') ? filePath : `${filePath}.zip`
 }
 
 /**
@@ -47,29 +32,6 @@ export function collectLogFiles(logDir: string): string[] {
         return false
       }
     })
-}
-
-async function createZipWithFiles(inputPaths: string[], outputZipPath: string): Promise<void> {
-  await fsPromises.mkdir(path.dirname(outputZipPath), { recursive: true })
-
-  await new Promise<void>((resolve, reject) => {
-    const zipFile = new yazl.ZipFile()
-    const output = fs.createWriteStream(outputZipPath)
-
-    const onError = (error: unknown): void => {
-      reject(error instanceof Error ? error : new Error(String(error)))
-    }
-
-    output.once('error', onError)
-    zipFile.outputStream.once('error', onError)
-    output.once('close', resolve)
-
-    zipFile.outputStream.pipe(output)
-    for (const inputPath of inputPaths) {
-      zipFile.addFile(inputPath, path.basename(inputPath))
-    }
-    zipFile.end()
-  })
 }
 
 /**
@@ -103,7 +65,7 @@ export async function exportLogsZip({
 
     const defaultPath = path.join(
       app.getPath('documents'),
-      buildDefaultLogsExportFilename(new Date()),
+      buildTimestampedZipName('memorylane-logs-export'),
     )
     const saveResult = await dialog.showSaveDialog(parentWindow ?? undefined, {
       title: 'Export Logs ZIP',
@@ -135,6 +97,3 @@ export async function exportLogsZip({
     return { success: false, error: message }
   }
 }
-
-// Exported for testing.
-export const __testing = { createZipWithFiles, buildDefaultLogsExportFilename, ensureZipExtension }
